Remove APICache
It seems that while caching is helpful for dealing with scenarios where the source site is down, it is much less helpful when the source site is updated and you need to "invalidate" the cache so that the client site displays the latest data. The client site itself seems pretty fast in deployment in scenarios where it must "reload" the data, so APICache itself is probably not worthwhile. Might as well remove it.